Revert "Move notification schemas and modules into central datastore module"
This reverts commit b33fad7861b610a8bf463b995b8d66d87998a494.
